:wrench: Improve Unit Tests for Utils
 - fix logger utility
